+++ packages/amplify-category-api/src/provider-utils/awscloudformation/cfn-api-artifact-handler.ts
@@ -12,7 +12,7 @@ import _ from 'lodash';
 import * as path from 'path';
 import uuid from 'uuid';
 import { category } from '../../category-constants';
-import { ApiArtifactHandler } from '../api-artifact-handler';
+import { ApiArtifactHandler, ApiArtifactHandlerOptions } from '../api-artifact-handler';
 import { cfnParametersFilename, gqlSchemaFilename, provider, rootAssetDir } from './aws-constants';
 import { authConfigHasApiKey, checkIfAuthExists, getAppSyncAuthConfig, getAppSyncResourceName } from './utils/amplify-meta-utils';
 import { appSyncAuthTypeToAuthConfig } from './utils/auth-config-to-app-sync-auth-type-bi-di-mapper';
@@ -99,7 +99,7 @@ class CfnApiArtifactHandler implements ApiArtifactHandler {
 
   // TODO once the AddApiRequest contains multiple services this class should depend on an ApiArtifactHandler
   // for each service and delegate to the correct one
-  updateArtifacts = async (request: UpdateApiRequest): Promise<void> => {
+  updateArtifacts = async (request: UpdateApiRequest, opts?: ApiArtifactHandlerOptions): Promise<void> => {
     const updates = request.serviceModification;
     const apiName = getAppSyncResourceName(this.context.amplify.getProjectMeta());
     if (!apiName) {
@@ -121,13 +121,16 @@ class CfnApiArtifactHandler implements ApiArtifactHandler {
     if (updates.additionalAuthTypes) {
       authConfig.additionalAuthenticationProviders = updates.additionalAuthTypes.map(appSyncAuthTypeToAuthConfig);
     }
+
     const logConfig = updates.logConfig;
-    await this.context.amplify.executeProviderUtils(this.context, 'awscloudformation', 'compileSchema', {
-      resourceDir,
-      parameters: this.getCfnParameters(apiName, authConfig, resourceDir),
-      authConfig,
-      logConfig,
-    });
+    if (!opts?.skipCompile) {
+      await this.context.amplify.executeProviderUtils(this.context, 'awscloudformation', 'compileSchema', {
+        resourceDir,
+        parameters: this.getCfnParameters(apiName, authConfig, resourceDir),
+        authConfig,
+        logConfig,
+      });
+    }
 
     this.context.amplify.updateamplifyMetaAfterResourceUpdate(category, apiName, 'output', { authConfig, logConfig });
     this.context.amplify.updateBackendConfigAfterResourceUpdate(category, apiName, 'output', { authConfig, logConfig });