+++ packages/amplify-headless-interface/src/interface/api/add.ts
@@ -148,6 +148,7 @@ export type AppSyncAuthType =
 export interface AppSyncAPIKeyAuthType {
   mode: 'API_KEY';
   expirationTime?: number;
+  apiKeyExpirationDate?: Date;
   keyDescription?: string;
 }