+++ packages/amplify-provider-awscloudformation/src/push-resources.ts
@@ -43,7 +43,7 @@ import { fileLogger } from './utils/aws-logger';
 import { APIGW_AUTH_STACK_LOGICAL_ID, loadApiWithPrivacyParams } from './utils/consolidate-apigw-policies';
 import { createEnvLevelConstructs } from './utils/env-level-constructs';
 import { NETWORK_STACK_LOGICAL_ID } from './network/stack';
-import { preProcessCFNTemplate } from './pre-push-cfn-processor/cfn-pre-processor';
+import { preProcessCFNTemplate, writeCustomPoliciesToCFNTemplate } from './pre-push-cfn-processor/cfn-pre-processor';
 import { AUTH_TRIGGER_STACK, AUTH_TRIGGER_TEMPLATE } from './utils/upload-auth-trigger-template';
 import { ensureValidFunctionModelDependencies } from './utils/remove-dependent-function';
 import { legacyLayerMigration, postPushLambdaLayerCleanup, prePushLambdaLayerPrompt } from './lambdaLayerInvocations';
@@ -73,14 +73,8 @@ export async function run(context: $TSContext, resourceDefinition: $TSObject) {
   let layerResources = [];
 
   try {
-    const {
-      resourcesToBeCreated,
-      resourcesToBeUpdated,
-      resourcesToBeSynced,
-      resourcesToBeDeleted,
-      tagsUpdated,
-      allResources,
-    } = resourceDefinition;
+    const { resourcesToBeCreated, resourcesToBeUpdated, resourcesToBeSynced, resourcesToBeDeleted, tagsUpdated, allResources } =
+      resourceDefinition;
     const cloudformationMeta = context.amplify.getProjectMeta().providers.awscloudformation;
     const {
       parameters: { options },
@@ -154,6 +148,7 @@ export async function run(context: $TSContext, resourceDefinition: $TSObject) {
     await transformGraphQLSchema(context, {
       handleMigration: opts => updateStackForAPIMigration(context, 'api', undefined, opts),
       minify: options['minify'],
+      promptApiKeyCreation: true,
     });
 
     // If there is a deployment already in progress we have to fail the push operation as another
@@ -395,13 +390,24 @@ export async function run(context: $TSContext, resourceDefinition: $TSObject) {
     if (iterativeDeploymentWasInvoked) {
       await deploymentStateManager.failDeployment();
     }
-    spinner.fail('An error occurred when pushing the resources to the cloud');
-
+    if(!await canAutoResolveGraphQLAuthError(error.message)) {
+      spinner.fail('An error occurred when pushing the resources to the cloud');
+    }
     rollbackLambdaLayers(layerResources);
 
     logger('run', [resourceDefinition])(error);
 
     throw error;
+  }
+}
+
+async function canAutoResolveGraphQLAuthError(message: string) {
+  if (message === `@auth directive with 'iam' provider found, but the project has no IAM authentication provider configured.`
+    || message === `@auth directive with 'userPools' provider found, but the project has no Cognito User Pools authentication provider configured.`
+    || message === `@auth directive with 'oidc' provider found, but the project has no OPENID_CONNECT authentication provider configured.`
+    || message === `@auth directive with 'apiKey' provider found, but the project has no API Key authentication provider configured.`
+    || message === `@auth directive with 'function' provider found, but the project has no Lambda authentication provider configured.`) {
+    return true;
   }
 }
 
@@ -759,10 +765,10 @@ function getCfnFiles(category: string, resourceName: string) {
 async function updateS3Templates(context: $TSContext, resourcesToBeUpdated: $TSAny, amplifyMeta: $TSMeta) {
   const promises = [];
 
-  for (const { category, resourceName } of resourcesToBeUpdated) {
+  for (const { category, resourceName, service } of resourcesToBeUpdated) {
     const { resourceDir, cfnFiles } = getCfnFiles(category, resourceName);
-
     for (const cfnFile of cfnFiles) {
+      await writeCustomPoliciesToCFNTemplate(resourceName, service, cfnFile, category);
       const transformedCFNPath = await preProcessCFNTemplate(path.join(resourceDir, cfnFile));
       promises.push(uploadTemplateToS3(context, transformedCFNPath, category, resourceName, amplifyMeta));
     }
@@ -833,7 +839,7 @@ async function formNestedStack(
       nestedStack.Description = 'Root Stack for AWS Amplify Console';
     }
   } catch (err) {
-    console.info('App not deployed yet.');
+    // if it is not an AmplifyAdmin app, do nothing
   }
 
   const { amplifyMeta } = projectDetails;
@@ -1018,6 +1024,11 @@ async function formNestedStack(
           }
         }
 
+        if ((category === 'api' || category === 'hosting') && resourceDetails.service === ApiServiceNameElasticContainer) {
+          parameters['deploymentBucketName'] = Fn.Ref('DeploymentBucketName');
+          parameters['rootStackName'] = Fn.Ref('AWS::StackName');
+        }
+
         const currentEnv = context.amplify.getEnvInfo().envName;
 
         if (!skipEnv && resourceName) {
@@ -1031,14 +1042,8 @@ async function formNestedStack(
         // If auth is imported check the parameters section of the nested template
         // and if it has auth or unauth role arn or name or userpool id, then inject it from the
         // imported auth resource's properties
-        const {
-          imported,
-          userPoolId,
-          authRoleArn,
-          authRoleName,
-          unauthRoleArn,
-          unauthRoleName,
-        } = context.amplify.getImportedAuthProperties(context);
+        const { imported, userPoolId, authRoleArn, authRoleName, unauthRoleArn, unauthRoleName } =
+          context.amplify.getImportedAuthProperties(context);
 
         if (category !== 'auth' && resourceDetails.service !== 'Cognito' && imported) {
           if (parameters.AuthCognitoUserPoolId) {