+++ packages/graphql-transformer-core/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.30.2](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformer-core@6.30.1...graphql-transformer-core@6.30.2) (2021-10-10)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+
+
+
+
+## [6.30.1](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformer-core@6.30.0...graphql-transformer-core@6.30.1) (2021-10-06)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+
+
+
+
 # [6.30.0](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformer-core@6.29.7...graphql-transformer-core@6.30.0) (2021-09-27)